requirements.txt
easyocr
Pillow
streamlit
pyarrow
